Upgrade dependencies and refactor backend models
