+++ backend/package.json
@@ -18,7 +18,7 @@
     "@types/uniqid": "^4.1.3",
     "@typescript-eslint/eslint-plugin": "^2.31.0",
     "@typescript-eslint/parser": "^2.31.0",
-    "axios": "^0.19.2",
+    "axios": "^0.21.1",
     "body-parser": "^1.19.0",
     "eslint": "^6.0.0",
     "eslint-config-prettier": "^6.11.0",
