--- backend/src/models/extendedSocket.ts
@@ -1,5 +1,0 @@
-import socketIo from "socket.io";
-
-export default interface ExtendedSocket extends socketIo.Socket {
-  username?: string;
-}
