--- backend/src/models/message.ts
@@ -1,4 +1,0 @@
-export default interface Message {
-  user?: string;
-  message: string;
-}
