--- backend/src/models/roomlist.ts
@@ -1,5 +1,0 @@
-import Room from "./room";
-
-export default interface RoomList {
-  [key: string]: Room;
-}
