--- backend/src/models/updateVideoStateRequest.ts
@@ -1,6 +1,0 @@
-import VideoState from "./videoState";
-
-export default interface UpdateVideoStateRequest {
-  videoState: VideoState;
-  socketId: string;
-}
